network: match version-buffer initialization and narrow packet tag
Full build08912eab succeeded13m05s. Empty aggregate initialization recovers the retail8B span; sign-in cur77.6519->79.6296 with18 spans and3 locals. Body change resets previous banked93.3481; historical headroom remains open. Explicit u8 removes C4244 with handle_send structure/score unchanged. Resolver comment cleanup byte-neutral. Global current report one improvement, no regressions or fold changes; remaining warnings recorded in campaign.

+++ sources/vostok/network_core/sources/udp_match_connection.cpp
@@ -93,7 +93,7 @@ void udp_match_connection::handle_send(
 
 	else {
 		pbyte	buffer	= packet->buffer_to_send( );
-		*buffer			= ( *pointer_cast< u16* >( buffer + 4 ) & 1 ) ? udp_match_multiple_packets : udp_match_single_packet;
+		*buffer			= u8( ( *pointer_cast< u16* >( buffer + 4 ) & 1 ) ? udp_match_multiple_packets : udp_match_single_packet );
 		m_unacknowledged_packets.push_back( packet );
 	}
